others/7008A_Clip-main/requirements.txt
sentence-transformers
torch
Pillow
numpy
tqdm
